refactor context.py module
