+++ context_logging/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqaF401
 from .context import Context, current_context, root_context
-from .setup_log_record import setup_log_record
+from .log_record import setup_log_record
 
 __version__ = '1.0.0'
