+++ context_logging/logger.py
@@ -1,0 +1,4 @@
+import logging
+
+logger = logging.getLogger(__package__)
+logger.addHandler(logging.NullHandler())
